Autotask-feature-Agent_email/requirements.txt
# Image Processing Dependencies
Pillow>=9.0.0
pytesseract>=0.3.10
opencv-python>=4.5.0
numpy>=1.21.0

# Automatic Processing Dependencies
schedule>=1.2.0
streamlit
pandas
plotly
pytz
schedule
snowflake-connector-python
dotenv
scikit-learn

# Google Calendar API Dependencies
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0